helix: disable mouse
I'm not ideologically opposed to using the mouse in the terminal.
But with a large scrolloff configured, clicking with the mouse into the
terminal can scroll the view, which is disorienting. With mouse disabled,
one can simply click into the terminal to return focus to it like one is used to.