hyprland: use zellij in scratch terminals
Hopefully it shouldn't pose any problems now that I'm not using a
plugin for the status bar anymore (which was buggy). I can always launch
a regular terminal without zellij normally. It happens kinda often to me
that I open a scratch terminal and then regret not being in zellij.