restructure user services
Instead of creating a symlink in the niri.service.wants/ directory for
every user service, there is only one entry that enables the custom
target `desktop-environment.target`. That target in turn lists all the
user services in the "Wants" section. I find this more maintainable.

It also means that if I ever remove a service from that file, the
removal itself will be managed by chezmoi. With creating and deleting
the symlink files, chezmoi will leave existing symlinks alone, even if
they are removed from the source directory. That means such services
would have to be deleted / disabled manually.