fix setting of SSH_AUTH_SOCK variable
Firstly, testing if it exists doesn't work (anymore?), because there is
some "keyring" agent running already. Secondly, I had to add the new
tests in the git cloning script as well, because when that is run at
first init time, the fish config is not loaded into the environment yet.