add bgi function ("buenzli git init")
This previously existed as a jj alias "jj bgi", which invoced a Rust
program. I removed it due to high complexity and low usage in:
b07394f57cfc42d8acb4a8898ce7558bca602416

As a shell function, it's much simpler and more powerful, because it can
modify the CWD directly.